[ide] Fix bug "declarationviewwindow empty"
On linux, declarationviewwindow is empty when triggered by
CompletionListWindow and potion=left.
In the OnExposeEvent of PopoverWindow, clipping is realized through a
DestIn operation which fails only with position=left.
I suspect a bug in cairo.
Disabling it does not seem to have bad effect on rendering and solve
the issue.